+++ src/main/java/com/sanaiclub/project/dao/ClientApplicantMapper.java
@@ -10,13 +10,13 @@ import java.util.List;
 public interface ClientApplicantMapper {
     List<ClientApplicantDTO> selectApplicantsByProjectId(@Param("projectId") Integer projectId);
 
-    ClientApplicantDTO selectApplicantDetail(@Param("applicationId") Long applicationId);
+    ClientApplicantDTO selectApplicantDetail(@Param("applicationId") Integer applicationId);
 
-    void updateStatus(@Param("applicationId") Long applicationId, @Param("status") String status);
+    void updateStatus(@Param("applicationId") Integer applicationId, @Param("status") String status);
 
-    List<String> selectFreelancerSkills(@Param("freelancerId") Long freelancerId);
+    List<String> selectFreelancerSkills(@Param("freelancerId") Integer freelancerId);
 
-    List<ClientApplicantDTO.CareerDTO> selectFreelancerCareers(@Param("freelancerId") Long freelancerId);
+    List<ClientApplicantDTO.CareerDTO> selectFreelancerCareers(@Param("freelancerId") Integer freelancerId);
 
-    List<ClientApplicantDTO.ProjectExpDTO> selectFreelancerProjectExps(@Param("freelancerId") Long freelancerId);
+    List<ClientApplicantDTO.ProjectExpDTO> selectFreelancerProjectExps(@Param("freelancerId") Integer freelancerId);
 }
